Separate eqdsk interface from parameterframe (bluemira#1456)
* Remove ParameterFrame from EQDSKInterface

* Make some EQDSKInterface attrs optional

Some of these attributes are 'extra' derived attributes that we've
deemed useful, so do not actually appear in EQDSK files (x, z, psinorm).
qpsi, doesn't seem to be in all EQDSK files.

* Update EQDSKInterface uses

* Tidy equilibria file test

* Fix 'psinorm'/'pnorm' confusion in EQDSKInterface

The file interface was reading psinorm in as 'psinorm', but writing it
out as 'pnorm'.

* Add EQDSKInterface test for derived values

* Remove mutable defaults from EQDSKInterface

* Resolve 'duplicated literal' sonarcloud warnings

* Map 'pnorm'->'psinorm' in EQDSKInterface reader

A bug in the reader meant that 'psinorm' was written out to 'pnorm'.
This bug has been fixed, but to keep backward compatibility with files
written whilst the bug was present, this change has been made.

* Tighten constraints in EQDSKInterface test

Co-authored-by: je-cook <81617086+je-cook@users.noreply.github.com>